feat: make the Rust compiler the only option
