--- packages/astro/src/core/compile/compile-rs.ts
@@ -1,155 +1,0 @@
-import { fileURLToPath } from 'node:url';
-import type { ResolvedConfig } from 'vite';
-import type { AstroConfig } from '../../types/public/config.js';
-import type { AstroError } from '../errors/errors.js';
-import { AggregateError, CompilerError } from '../errors/errors.js';
-import { AstroErrorData } from '../errors/index.js';
-import { normalizePath, resolvePath } from '../viteUtils.js';
-import { createStylePreprocessor, type PartialCompileCssResult } from './style.js';
-import type { CompileCssResult } from './types.js';
-
-export interface CompileProps {
-	astroConfig: AstroConfig;
-	viteConfig: ResolvedConfig;
-	toolbarEnabled: boolean;
-	filename: string;
-	source: string;
-}
-
-export interface CompileResult {
-	code: string;
-	map: string;
-	scope: string;
-	css: CompileCssResult[];
-	scripts: any[];
-	hydratedComponents: any[];
-	clientOnlyComponents: any[];
-	serverComponents: any[];
-	containsHead: boolean;
-	propagation: boolean;
-	styleError: string[];
-	diagnostics: any[];
-}
-
-export async function compile({
-	astroConfig,
-	viteConfig,
-	toolbarEnabled,
-	filename,
-	source,
-}: CompileProps): Promise<CompileResult> {
-	let preprocessStyles;
-	let transform;
-	try {
-		({ preprocessStyles, transform } = await import('@astrojs/compiler-rs'));
-	} catch (err: unknown) {
-		throw new Error(
-			`Failed to load @astrojs/compiler-rs. Make sure it is installed and up to date. Original error: ${err}`,
-		);
-	}
-
-	const cssPartialCompileResults: PartialCompileCssResult[] = [];
-	const cssTransformErrors: AstroError[] = [];
-	let transformResult: any;
-
-	try {
-		const preprocessedStyles = await preprocessStyles(
-			source,
-			createStylePreprocessor({
-				filename,
-				viteConfig,
-				astroConfig,
-				cssPartialCompileResults,
-				cssTransformErrors,
-			}),
-		);
-
-		transformResult = transform(source, {
-			compact: astroConfig.compressHTML,
-			filename,
-			normalizedFilename: normalizeFilename(filename, astroConfig.root),
-			sourcemap: 'both',
-			internalURL: 'astro/compiler-runtime',
-			// TODO: remove in Astro v7
-			astroGlobalArgs: JSON.stringify(astroConfig.site),
-			scopedStyleStrategy: astroConfig.scopedStyleStrategy,
-			resultScopedSlot: true,
-			transitionsAnimationURL: 'astro/components/viewtransitions.css',
-			annotateSourceFile:
-				viteConfig.command === 'serve' &&
-				astroConfig.devToolbar &&
-				astroConfig.devToolbar.enabled &&
-				toolbarEnabled,
-			preprocessedStyles,
-			resolvePath(specifier) {
-				return resolvePath(specifier, filename);
-			},
-		});
-	} catch (err: any) {
-		// The compiler should be able to handle errors by itself, however
-		// for the rare cases where it can't let's directly throw here with as much info as possible
-		throw new CompilerError({
-			...AstroErrorData.UnknownCompilerError,
-			message: err.message ?? 'Unknown compiler error',
-			stack: err.stack,
-			location: {
-				file: filename,
-			},
-		});
-	}
-
-	handleCompileResultErrors(filename, transformResult, cssTransformErrors);
-
-	return {
-		...transformResult,
-		css: transformResult.css.map((code: string, i: number) => ({
-			...cssPartialCompileResults[i],
-			code,
-		})),
-	};
-}
-
-function handleCompileResultErrors(
-	filename: string,
-	result: any,
-	cssTransformErrors: AstroError[],
-) {
-	const compilerError = result.diagnostics.find((diag: any) => diag.severity === 'error');
-
-	if (compilerError) {
-		throw new CompilerError({
-			name: 'CompilerError',
-			message: compilerError.text,
-			location: {
-				line: compilerError.labels[0].line,
-				column: compilerError.labels[0].column,
-				file: filename,
-			},
-			hint: compilerError.hint,
-		});
-	}
-
-	switch (cssTransformErrors.length) {
-		case 0:
-			break;
-		case 1: {
-			throw cssTransformErrors[0];
-		}
-		default: {
-			throw new AggregateError({
-				...cssTransformErrors[0],
-				errors: cssTransformErrors,
-			});
-		}
-	}
-}
-
-function normalizeFilename(filename: string, root: URL) {
-	const normalizedFilename = normalizePath(filename);
-	const normalizedRoot = normalizePath(fileURLToPath(root));
-	if (normalizedFilename.startsWith(normalizedRoot)) {
-		return normalizedFilename.slice(normalizedRoot.length - 1);
-	} else {
-		return normalizedFilename;
-	}
-}
