+++ packages/astro/src/core/compile/compile.ts
@@ -1,6 +1,9 @@
 import { fileURLToPath } from 'node:url';
-import type { TransformResult } from '@astrojs/compiler';
-import { transform } from '@astrojs/compiler';
+import {
+	preprocessStyles,
+	transform,
+	type TransformResult,
+} from '@astrojs/compiler-rs';
 import type { ResolvedConfig } from 'vite';
 import type { AstroConfig } from '../../types/public/config.js';
 import type { AstroError } from '../errors/errors.js';
@@ -29,18 +32,23 @@ export async function compile({
 	filename,
 	source,
 }: CompileProps): Promise<CompileResult> {
-	// Because `@astrojs/compiler` can't return the dependencies for each style transformed,
-	// we need to use an external array to track the dependencies whenever preprocessing is called,
-	// and we'll rebuild the final `css` result after transformation.
 	const cssPartialCompileResults: PartialCompileCssResult[] = [];
 	const cssTransformErrors: AstroError[] = [];
 	let transformResult: TransformResult;
 
 	try {
-		// Transform from `.astro` to valid `.ts`
-		// use `sourcemap: "both"` so that sourcemap is included in the code
-		// result passed to esbuild, but also available in the catch handler.
-		transformResult = await transform(source, {
+		const preprocessedStyles = await preprocessStyles(
+			source,
+			createStylePreprocessor({
+				filename,
+				viteConfig,
+				astroConfig,
+				cssPartialCompileResults,
+				cssTransformErrors,
+			}),
+		);
+
+		transformResult = transform(source, {
 			compact: astroConfig.compressHTML,
 			filename,
 			normalizedFilename: normalizeFilename(filename, astroConfig.root),
@@ -56,14 +64,8 @@ export async function compile({
 				astroConfig.devToolbar &&
 				astroConfig.devToolbar.enabled &&
 				toolbarEnabled,
-			preprocessStyle: createStylePreprocessor({
-				filename,
-				viteConfig,
-				astroConfig,
-				cssPartialCompileResults,
-				cssTransformErrors,
-			}),
-			async resolvePath(specifier) {
+			preprocessedStyles,
+			resolvePath(specifier) {
 				return resolvePath(specifier, filename);
 			},
 		});
@@ -80,7 +82,7 @@ export async function compile({
 		});
 	}
 
-	handleCompileResultErrors(transformResult, cssTransformErrors);
+	handleCompileResultErrors(filename, transformResult, cssTransformErrors);
 
 	return {
 		...transformResult,
@@ -91,19 +93,21 @@ export async function compile({
 	};
 }
 
-function handleCompileResultErrors(result: TransformResult, cssTransformErrors: AstroError[]) {
-	// TODO: Export the DiagnosticSeverity enum from @astrojs/compiler?
-	// eslint-disable-next-line @typescript-eslint/no-unsafe-enum-comparison
-	const compilerError = result.diagnostics.find((diag) => diag.severity === 1);
+function handleCompileResultErrors(
+	filename: string,
+	result: TransformResult,
+	cssTransformErrors: AstroError[],
+) {
+	const compilerError = result.diagnostics.find((diag) => diag.severity === 'error');
 
 	if (compilerError) {
 		throw new CompilerError({
 			name: 'CompilerError',
 			message: compilerError.text,
 			location: {
-				line: compilerError.location.line,
-				column: compilerError.location.column,
-				file: compilerError.location.file,
+				line: compilerError.labels[0].line,
+				column: compilerError.labels[0].column,
+				file: filename,
 			},
 			hint: compilerError.hint,
 		});
